new try

+++ hashmap.c
@@ -46,7 +46,7 @@ int is_equal(void* key1, void* key2){
 
 
 void insertMap(HashMap * map, char * key, void * value) {
-    int pos = hash(key, map->capacity);
+    unsigned long pos = hash(key, map->capacity);
     int aux = pos;
     bool libre = true;
     struct Pair *n = createPair(key, value);
